+++ packages/orchestration/src/examples/sendAnywhere.contract.js
@@ -1,13 +1,12 @@
 import { makeStateRecord } from '@agoric/async-flow';
 import { AmountShape } from '@agoric/ertp';
-import { heapVowE } from '@agoric/vow/vat.js';
 import { InvitationShape } from '@agoric/zoe/src/typeGuards.js';
 import { Fail } from '@endo/errors';
 import { E } from '@endo/far';
 import { M } from '@endo/patterns';
-import { CosmosChainInfoShape } from '../typeGuards.js';
 import { withOrchestration } from '../utils/start-helper.js';
 import { orchestrationFns } from './sendAnywhereFlows.js';
+import { prepareChainHubAdmin } from '../exos/chain-hub-admin.js';
 
 /**
  * @import {TimerService} from '@agoric/time';
@@ -16,7 +15,6 @@ import { orchestrationFns } from './sendAnywhereFlows.js';
  * @import {Remote, Vow} from '@agoric/vow';
  * @import {Zone} from '@agoric/zone';
  * @import {VBankAssetDetail} from '@agoric/vats/tools/board-utils.js';
- * @import {CosmosChainInfo, IBCConnectionInfo} from '../cosmos-api';
  * @import {CosmosInterchainService} from '../exos/cosmos-interchain-service.js';
  * @import {OrchestrationTools} from '../utils/start-helper.js';
  */
@@ -60,6 +58,8 @@ const contract = async (
     },
   );
 
+  const creatorFacet = prepareChainHubAdmin(zone, chainHub);
+
   // TODO should be a provided helper
   /** @type {(brand: Brand) => Vow<VBankAssetDetail>} */
   const findBrandInVBank = vowTools.retriable(
@@ -96,38 +96,6 @@ const contract = async (
           undefined,
           M.splitRecord({ give: SingleAmountRecord }),
         );
-      },
-    },
-  );
-
-  let nonce = 0n;
-  const ConnectionInfoShape = M.record(); // TODO
-  const creatorFacet = zone.exo(
-    'Send CF',
-    M.interface('Send CF', {
-      addChain: M.callWhen(CosmosChainInfoShape, ConnectionInfoShape).returns(
-        M.scalar(),
-      ),
-    }),
-    {
-      /**
-       * @param {CosmosChainInfo} chainInfo
-       * @param {IBCConnectionInfo} connectionInfo
-       */
-      async addChain(chainInfo, connectionInfo) {
-        const chainKey = `${chainInfo.chainId}-${(nonce += 1n)}`;
-        // when() because chainHub methods return vows. If this were inside
-        // orchestrate() the membrane would wrap/unwrap automatically.
-        const agoricChainInfo = await heapVowE.when(
-          chainHub.getChainInfo('agoric'),
-        );
-        chainHub.registerChain(chainKey, chainInfo);
-        chainHub.registerConnection(
-          agoricChainInfo.chainId,
-          chainInfo.chainId,
-          connectionInfo,
-        );
-        return chainKey;
       },
     },
   );
