+++ packages/orchestration/test/examples/snapshots/sendAnywhere.test.ts.md
@@ -25,8 +25,8 @@ Generated by [AVA](https://avajs.dev).
         unwrapMap: 'Alleged: weakMapStore',
       },
       contract: {
-        'Send CF_kindHandle': 'Alleged: kind',
-        'Send CF_singleton': 'Alleged: Send CF',
+        'ChainHub Admin_kindHandle': 'Alleged: kind',
+        'ChainHub Admin_singleton': 'Alleged: ChainHub Admin',
         'Send PF_kindHandle': 'Alleged: kind',
         'Send PF_singleton': 'Alleged: Send PF',
       },
